added general abstract button
